+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for mason
 
+## Unreleased
+
+* Added `doubleFixed`, `doubleSI` and `doubleExp`
+
 ## 0 -- 2019-12-05
 
 * First version. Released on an unsuspecting world.